modifying arg of function

+++ 0x14-bit_manipulation/holberton.h
@@ -2,7 +2,7 @@
 #define HOLBERTON_H
 
 unsigned int binary_to_uint(const char *);
-void print_binary(unsigned int);
+void print_binary(unsigned long int);
 int _putchar(char);
 
 #endif
